Implement partial recursive resolver
This kind of works, but has one small, one large, and at least one
unknown problem.

The small problem: it only uses TCP to reach out to other nameservers.
It should use UDP as well.

The large problem: it only works if all the nameservers from the root
downwards return both AUTHORITY and ADDITIONAL sections with the NS
records in the AUTHORITY and the A records for those nameservers in
the ADDITIONAL.  This works fine for .com, but not for .uk.  A more
general solution will also need to handle NS records in the ANSWER,
and to recursively resolve hostnames in NS records which don't have an
A record in the same response.  Care will have to be taken so that bad
configuration can't cause an infinite loop or something - maybe just a
request count limit?

The unknown problem: `dig www.barrucadu.com` doesn't work.  That
domain is a CNAME to `barrucadu.com`, am I mishandling CNAMEs
somewhere?  Probably.