Add deserialisation round-trip fuzz test with `cargo fuzz`
This test looks for crashes in the deserialisation logic.  It didn't
manage to find any - but given that there are no panics in there and
all slice indexing is checked, maybe that's not too surprising.

If I inject subtle bugs - like a panic if a domain name has a pointer,
pointing to something 10 octets long, it does find it trivially.  So
it does seem to generate valid messages with interesting data pretty
easily, which is good to see.

It would have been nice if this found a bug, but I think it's somewhat
reassuring even without that.