Make cache actually an LRU cache
I was pruning by access count, which is possibly a good idea - if
something is accessed a lot, it's worth keeping around - but also
penalises new entries, when there is also an argument that if
something has just been requested, it is likely to be requested again
in the near future.

Since LRU is a standard cache eviction approach, it's probably better
overall.  So instead of tracking and evicting by access count, use
access time.