Implement zones & migrate static records to zones
How I hoped to avoid needing to implement proper zones; but it looks
like it really is the best way to support some of the later DNS RFCs.

If I'd just been more forward-thinking in the beginning, I'd have
started with a zone structure from the get-go.  Now I need to shoehorn
it in and migrate the rest of the code over, oh well.