Rewrite nonrecursive resolver to more closely match the RFC
Much of the logic moves into the `Zone` type, leaving
`resolve_nonrecursive` the task of appropriately handling
non-authoritative zones, expanding CNAMEs, and querying the cache
where an authoritative zone cannot answer.

This commit also makes the resolver return:

- a SOA record if the query is answered by an authoritative zone and
is not a CNAME (for simplicify and brevity - returning all the
relevant SOAs for CNAMEs which cross zone boundaries feels like a
pain).

- a delegation if the query is answered by an authoritative zone which
delegates a superdomain of the question domain.

- a NameError response if there is only one top-level query, it is
answered by an authoritative zone, and the name does not exist.

Implementing delegation needed me to make a decision about the
semantics of wildcard NS records.

Say we have these NS records:

    *.example.com    NS    ns1.example.net

And receive a query for `long.subdomain.of.example.com`.  What gets
returned?

With any other wildcard record, the full domain gets substituted in.
But should that happen here, or should only the topmost label get
substituted in?  Which of these is correct?

    long.subdomain.of.example.com    NS    ns1.example.net

Or

    of.example.com    NS    ns1.example.net

Unfortunately, RFC 4592, which clarifies the use of wildcard RRs,
says:

> there is no definition given for wildcard domain names owning an NS
> RRSet [...] inclusion of wildcard NS RRSets in a zone is
> discouraged, but not barred.

So I decided to use the second option, of only including the next
label down.  I've not checked what other nameservers do, but this
seems sensible enough.