Enforce domain names are ASCII only
RFC 1035 is a bit inconsistent with character encoding of domain
names.  On the one hand, it says you *shouldn't* assume ASCII:

> For all parts of the DNS that are part of the official protocol, all
> comparisons between character strings (e.g., labels, domain names,
> etc.)  are done in a case-insensitive manner.  At present, this rule
> is in force throughout the domain system without exception.
> However, future additions beyond current usage may need to use the
> full binary octet capabilities in names, so attempts to store domain
> names in 7-bit ASCII or use of special bytes to terminate labels,
> etc., should be avoided.

But on the other hand, it says you *should* assume ASCII for the
case-insensitivity:

> Name servers and resolvers must compare labels in a case-insensitive
> manner (i.e., A=a), assuming ASCII with zero parity.  Non-alphabetic
> codes must match exactly.

I don't think it's sensible to both assume domain names can be
arbitrary bytes and also treat those arbitrary bytes as ASCII for the
purposes of case-folding.  That would lead to incorrect case-folding
of non-ASCII-encoded names!  And punycode (RFC 3492) encodes unicode
to ASCII.

So, in practice, I think assuming ASCII for domain names is safe, and
may actually be guaranteed by RFC 3492 (which I've only read the
abstract of).