Fix non-ASCII zone file shenanigans
To avoid character encoding issues, any non-ASCII bytes should be
escaped in the numeric form.

Furthermore, this:

    "\DDD" -> char -> u8

is not a safe conversion, as if that `\DDD` is some unicode magic it
may give rise to a char which converts into a different u8.

So pass around the u8s and the chars.  The chars are convenient for
most of the parsing, the u8s are needed for the record types which
hold uninterpreted octets.