Reduce TTL of hosts records to 5s
This is to improve the responsiveness of reloading configuration.  If
something is removed, and there's a caching resolver downstream, we
don't want to have to wait 5 minutes for that cached entry to expire.

Resolving static records is really fast, even with large zones, so
there's little harm in these queries hitting resolved more frequently.