Model non-recursive resolver as a pseudo-nameserver
This commit makes the non-recursive resolver return a
`NameserverResponse`, just like an upstream nameserver which the
recursive resolver may call normally.  This makes it straightforward
to properly handle delegations and unresolved CNAMEs in the recursive
resolver by pattern-matching on the result and not needing to
interpret the RRs.

I chose this approach, rather than defining yet another
wrapper-around-RRs type because `NameserverResponse` already covers
delegation and CNAMEs, and so just needed to be extended with a notion
of authority in some places.

Furthermore, re-using this type feels cleaner than adding to the
`ResolvedRecord` type, which is supposed to be a direct representation
of a response to send to the client.

Closes #48