Make non-authoritative records override cached ones
If a non-wildcard query has results in a non-authoritative zone,
return just those RRs.  Furthermore, for wildcard queries, only
include RRs from the cache if there aren't any of those types in the
zone.

This makes non-authoritative zones act as an override on top of the
rest of the domain name system.  Which is essential to avoid weird
inconsistent behaviour like:

1. A query is made and some RRs cached
2. The config is updated to override that query
3. The query is repeated, and both the cached and non-authoritative
   records are merged together.

Closes #65