Ignore interface-specific addresses in hosts files
StevenBlack/hosts is a widely used blocklist which has one weird entry
in it which tripped up my parser:

    fe80::1%lo0 localhost

This certainly looks like an interface-specific address.  I don't know
what that would mean, though, and it's not described in hosts(5).  So
I'm just going to ignore such entries, to improve compatibility with
existing blocklists.

Closes #63