Use `tracing` crate for logging
The following levels are used in `dns-resolver`:

- `TRACE`: logs on most function calls and match branches, this more
or less replicates the previous logging

- `DEBUG`: logs on suspicious but minor errors which indicate
something is probably wrong externally

- `WARN`: logs on non-fatal internal errors, but these probably
indicate bugs.

- `ERROR`: logs on fatal errors and then immediately exits.

The only fatal errors in `dns-resolver` are in the network utility
functions.

The following levels are used in `resolved`:

- `INFO`: logs top-level events, like new connections or answered
questions

- `WARN`: logs on non-fatal internal and configuration errors.

- `ERROR`: logs on fatal errors and then immediately exits.

Log level can be configured with the `RUST_LOG` env var.  See the
README for an example.

Closes #19