Prevent cycles in questions
If a domain's nameservers are incorrectly configured such that `A`
records can't be found (eg, missing glue records), the recursive
resolver will bounce back and forth between the nameservers until it
hits the recursion limit.  If the domain has several nameservers, this
may result in a lot of waiting while many different redundant orderings
get tried.

However, we can bail out much sooner if we detect a cycle: if finding
the `A` record for `ns1.example.com` requires (even if indirectly)
finding the `A` record for `ns1.example.com`, then we're stuck.  There's
no point going around this cycle 25 times until the recursion depth is
hit.  We can abort then.

So this commit replaces the `recursion_limit` with a `question_stack`,
and resolvers now check if they've been caught in a loop in addition to
checking the recursion depth.