When possible, use a glue record as the answer
EasyDNS is a DNS provider.  The EasyDNS nameservers are:

    dns1.easydns.com.
    dns2.easydns.net.
    dns3.easydns.org.
    dns4.easydns.info.

Note two things:

1. These are all subdomains of EasyDNS, which means they will have glue
records: `A` records which the registry holds, to point to the `NS`
records, otherwise we have a bootstrapping problem where to resolve
`easydns.com.` you must first resolve `easydns.com.`.

2. These are all different TLDs, presumably so that if `com.`, `net.`,
and `org.` simultaneously go down people can still use EasyDNS - even
though we're now in the middle of a nuclear war (what else could knock
out three major TLDs at the same time?).

If we resolve `easydns.com.` against the `com.` nameserver, we get:

    ; <<>> DiG 9.18.7 <<>> @a.gtld-servers.net. easydns.com
    ; (2 servers found)
    ;; global options: +cmd
    ;; Got answer:
    ;; ->>HEADER<<- opcode: QUERY, status: NOERROR, id: 36389
    ;; flags: qr rd; QUERY: 1, ANSWER: 0, AUTHORITY: 4, ADDITIONAL: 3
    ;; WARNING: recursion requested but not available

    ;; OPT PSEUDOSECTION:
    ; EDNS: version: 0, flags:; udp: 4096
    ;; QUESTION SECTION:
    ;easydns.com.                   IN      A

    ;; AUTHORITY SECTION:
    easydns.com.            172800  IN      NS      dns3.easydns.org.
    easydns.com.            172800  IN      NS      dns1.easydns.com.
    easydns.com.            172800  IN      NS      dns2.easydns.net.
    easydns.com.            172800  IN      NS      dns4.easydns.info.

    ;; ADDITIONAL SECTION:
    dns1.easydns.com.       172800  IN      AAAA    2400:cb00:2049:1::a29f:1835
    dns1.easydns.com.       172800  IN      A       64.68.192.10

    ;; Query time: 33 msec
    ;; SERVER: 192.5.6.30#53(a.gtld-servers.net.) (UDP)
    ;; WHEN: Fri Nov 18 17:56:11 GMT 2022
    ;; MSG SIZE  rcvd: 194

That is, all of the `NS` records (since `easydns.com.` is a subdomain of
`com.`) but the glue record for only `dns1.easydns.com.` - because the
other nameservers are NOT subdomains of `easydns.com.`

This is correct behaviour.

Unfortunately, resolved didn't handle "incomplete" sets of glue records
like this very well.  Say we're resolving `easydns.com.` and get that
response.  resolved caches all those records and now knows it needs to
query one of those four nameservers.  But which nameserver it picks is
essentially arbitrary: it has the `A` record for `dns1.easydns.com.` in
its cache, but may not pick `dns1.easydns.com.` to answer the query.  In
fact, since there are 4 nameservers there, it's got a 3/4 chance of
picking a different one - one it doesn't have the `A` record for.

And so if it picks, say, `dns2.easydns.net.` it ends up resolving
`easydns.net.` and getting in the same situation: now it has a 2/4
chance of picking a nameserver it has an `A` record for.

And so on.

But all those queries are unnecessary!  It should have just used the
glue record for `dns1.easydns.com.` in the first place, rather than
trying to go to another nameserver.

So, change the delegation handler to check if we got back a glue record
which answers the current query: if we did, great, return that;
otherwise proceed as before.