Prefer nameservers which can be resolved locally
When choosing from the candidate nameservers, skip any which can't be
resolved locally.  If no candidates can be resolved locally, restart
and start resolving them remotely.

This is useful if we have incomplete sets of glue records.  For example,
if we resolve `easydns.com.` against the `com.` nameserver, we get:

    ; <<>> DiG 9.18.7 <<>> @a.gtld-servers.net. easydns.com
    ; (2 servers found)
    ;; global options: +cmd
    ;; Got answer:
    ;; ->>HEADER<<- opcode: QUERY, status: NOERROR, id: 36389
    ;; flags: qr rd; QUERY: 1, ANSWER: 0, AUTHORITY: 4, ADDITIONAL: 3
    ;; WARNING: recursion requested but not available

    ;; OPT PSEUDOSECTION:
    ; EDNS: version: 0, flags:; udp: 4096
    ;; QUESTION SECTION:
    ;easydns.com.                   IN      A

    ;; AUTHORITY SECTION:
    easydns.com.            172800  IN      NS      dns3.easydns.org.
    easydns.com.            172800  IN      NS      dns1.easydns.com.
    easydns.com.            172800  IN      NS      dns2.easydns.net.
    easydns.com.            172800  IN      NS      dns4.easydns.info.

    ;; ADDITIONAL SECTION:
    dns1.easydns.com.       172800  IN      AAAA    2400:cb00:2049:1::a29f:1835
    dns1.easydns.com.       172800  IN      A       64.68.192.10

    ;; Query time: 33 msec
    ;; SERVER: 192.5.6.30#53(a.gtld-servers.net.) (UDP)
    ;; WHEN: Fri Nov 18 17:56:11 GMT 2022
    ;; MSG SIZE  rcvd: 194

We get a delegation of four nameservers, but glue records for only one
of them (this is correct since the other nameservers are not subdomains
of `easydns.com.`).

The current behaviour is to update the list of candidates, restart, and
pick essentially an arbitrary one (since the order of records returned
from the upstream nameserver is arbitrary).  But instead we should pick
`dns1.easydns.com.`, since we now have its `A` record cached.

This is similar to b77f8f6, but solves a different problem.  Both remove
avoidable queries.

b77f8f6 fixes the case where we try to resolve the `A` address of the
nameserver (`dns1.easydns.com.`) and get an incomplete set of glue
records: prior to b77f8f6 we would have asked one of the delegated
nameservers for the `A` record for `dns1.easydns.com` - which is
redundant since we have the glue record.

Fixes #199