Fix potential DoS with responding to responses
An attacker could send a packet with the source address and port being
resolved's, which will make resolved send the response to itself.  But
resolved responds to an unexpected response by sending a format
error *response*.  Which would go back to itself, again, triggering an
infinite loop of responses.

Much as it pains me to just drop a perfectly decodable, but invalid,
message, this is the best solution since there isn't really a way to
reliably tell if an attacker is triggering this case (eg, they could use
two instances running on different IPs and bounce responses between
them).

See #246