Refuse queries with `QDCOUNT > 1`
This appears to be pretty common practice, and means I don't need to
reason about how to handle these requests; eg, the case where there's a
mixture of answerable and unanswerable questions in the same query goes
away.

Also add a metric to track refused queries and the reason why.