Return a single SOA record in authoritative responses
This refactors the `ResolvedRecord` type to have a single (required)
`soa_rr`, instead of a possibly empty collection of `authority_rrs`,
which has knock-on effects in the local resolver.

Previously the local resolver could return an authoritative response
with no SOA, which doesn't make much sense.  So, in those cases,
propagate the SOA RR of the authoritative zone that answered the
question - which is what it should have been doing anyway.

With this change, there's no longer any obvious way to return NS RRs.
But I don't return those currently, and if I decide to start doing so, I
can always add a new field.