Propagate SOA record for NXDOMAIN / NODATA responses
This makes the recursive & forwarding resolvers propagate the SOA record
from upstream NXDOMAIN / NODATA responses, so that resolved won't return
those as ServFail any more.

This does some basic validation of the SOA to make sure we're not
passing along nonsense, but there's some trust involved (especially in
forwarding mode).

This does *not* make resolved return an NXDOMAIN result for any domain
it's not authoritative for.  Interestingly, Cloudflare and Google DNS
do:

```
; <<>> DiG 9.18.16 <<>> @1.1.1.1 google.c-m
; (1 server found)
;; global options: +cmd
;; Got answer:
;; ->>HEADER<<- opcode: QUERY, status: NXDOMAIN, id: 60891
;; flags: qr aa rd ra; QUERY: 1, ANSWER: 0, AUTHORITY: 1, ADDITIONAL: 1

;; OPT PSEUDOSECTION:
; EDNS: version: 0, flags:; udp: 1232
;; QUESTION SECTION:
;google.c-m.                    IN      A

;; AUTHORITY SECTION:
.                       86400   IN      SOA     a.root-servers.net. nstld.verisign-grs.com. 2023092100 1800 900 604800 86400

;; Query time: 12 msec
;; SERVER: 1.1.1.1#53(1.1.1.1) (UDP)
;; WHEN: Sat Sep 23 23:20:30 BST 2023
;; MSG SIZE  rcvd: 114
```

My reading of RFC 1035 suggests that that's not quite valid, as only an
authoritative nameserver should return an NXDOMAIN response.  But,
curiously, the response has the AA flag set.  Maybe they're just
propagating the authority in the case of NXDOMAIN responses?

resolved will just supply a NODATA response in that case:

```
; <<>> DiG 9.18.16 <<>> @127.0.0.1 google.c-m
; (1 server found)
;; global options: +cmd
;; Got answer:
;; ->>HEADER<<- opcode: QUERY, status: NOERROR, id: 31834
;; flags: qr rd ra; QUERY: 1, ANSWER: 0, AUTHORITY: 1, ADDITIONAL: 0

;; QUESTION SECTION:
;google.c-m.                    IN      A

;; AUTHORITY SECTION:
.                       86400   IN      SOA     a.root-servers.net. nstld.verisign-grs.com. 2023092301 1800 900 604800 86400

;; Query time: 23 msec
;; SERVER: 127.0.0.1#53(127.0.0.1) (UDP)
;; WHEN: Sat Sep 23 23:22:33 BST 2023
;; MSG SIZE  rcvd: 103
```

Maybe I could return an NXDOMAIN here?  Anyway, this is better than it
was before.

Closes #248