Allow forwarding queries via IPv6 and to any port
This changes the `--forward-address` arg of resolved and dnsq to take an
arbitrary socket address (a string in the form `ip:port`) rather than
assuming it's an IPv4 address on port 53.  This does mean the port is
now required, but I think that's a small price to pay for the extra
flexibility.