Allow listening on IPv6 and any port
This changes the `--interface` and `--metrics-interface` args of
resolved to take an arbitrary socket address (a string in the form
`ip:port`) rather than assuming it's an IPv4 address on port 53 (for the
`interface`) or on `--metrics-port` (for the `metrics-interface`).

This is a slight breaking change as it does mean the port is now
requires, and the `--metrics-port` argument is no more, but I think the
extra flexibility is worth it.