Add in-principle support for querying nameservers over IPv6
This adds a new flag `--protocol-mode` which takes one of four values:

- `only-v4` - only use IPv4 (e.g. this is an IPv4-only network),
rejecting a nameserver if it is only available over IPv6.

- `prefer-v4` - if a nameserver is only available over IPv6, use that;
but if it is available over both IPv4 and IPv6 (or only IPv4), use the
IPv4 address.

- `prefer-v6` - if a nameserver is only available over IPv4, use that;
but if it is available over both IPv4 and IPv6 (or only IPv6), use the
IPv6 address.

- `only-v6` - only use IPv6 (e.g. this is an IPv6-only network),
rejecting a nameserver if it is only available over IPv4.

I only have IPv4 at home, so this isn't really tested, but it should
work.  The default is `only-v4`.

Right now resolved suffers from the problem of assuming one hostname
maps to one IP address.  So if a nameserver has both A and AAAA records,
and the protocol mode is set to a `prefer-*` option, but that address
doesn't work, it won't try the other one.  For that matter, if there are
multiple A (or AAAA) records and the first one it tries doesn't work, it
won't try the others.  This hasn't caused any problems that I've
noticed, but it would be nice to fix.

This commit also fixes the case where a glue AAAA record answers the
query, and so can be used to avoid an additional nameserver query.
Previously only glue A records were used for that.