Implement whole-(Q)NAME compression
Compress names in QNAME or NAME position (ie, the name a question or
record refers to, but not a name in the record data) if an exact copy of
that name has been seen before.

New names are only introduced as a QNAME or in RDATA, and while the
RDATA names aren't being compressed they are being used to compress
later names.  So in practice, this should mean that all names in (Q)NAME
position (except for the very first question with that QNAME) should be
compressed.

This commit only implements whole-name compression, whereas suffixes can
be compressed too.  That would be more complex, and I think we'll get
most of the benefit of compression just from whole names anyway.