Add basic documentation site
This is really minimal, just some README tweaks to make it read better
as a standalone webpage (while still working as a README) and a script
to render it with mdbook.

Future work will be to write some more documentation pages (eg, a more
detailed treatment of configuration, or how DNS works) and to review /
rewrite the rustdocs.