Refactor: split PartitionedCache out of Cache
There's a more generic cache data structure hiding inside `Cache`, and
hopefully separating the RR-specific part from the data structure part
makes it easier to understand (and easier to change the implementation
of).  I've left the tests as they are, against the `Cache` rather than
the new `PartitionedCache`, but I might change that in the future.